// src/app/interfaces/amigos.tsx
export interface Amigo {
  id_promovido: string;
  nombre: string;
  primer_apellido: string;
  segundo_apellido: string;
  ciudad: string;
  colonia: string;
  calle: string;
  numero: string;
  clave_electoral: string;
  curp: string;
  entidad: string;
  seccion_electoral: string;
  correo: string;
  telefono: string;
  isActive: boolean;
  promoventeInmediatoRol: string;
  createdAt: string;
  structureRol: string;
  voto: boolean;
}